# Performance Profiler: updates loading screen.