--- a/client/performance-profiler/pages/dashboard/index.tsx
+++ b/client/performance-profiler/pages/dashboard/index.tsx
@@ -34,6 +34,7 @@ export const PerformanceProfilerDashboard = ( props: PerformanceProfilerDashboar
 	return (
 		<div className="container">
 			<DocumentHead title={ translate( 'Speed Test' ) } />
+
 			<PerformanceProfilerHeader
 				url={ url }
 				activeTab={ activeTab }