--- a/client/performance-profiler/pages/dashboard/style.scss
+++ b/client/performance-profiler/pages/dashboard/style.scss
@@ -1,5 +1,6 @@
 .is-group-performance-profiler .layout__content {
 	padding: 0;
+	background: #fff;
 }
 
 .l-block-wrapper {