--- a/client/performance-profiler/pages/dashboard/style.scss
+++ b/client/performance-profiler/pages/dashboard/style.scss
@@ -7,3 +7,13 @@
 	margin: 0 auto;
 	max-width: 1056px;
 }
+
+.peformance-profiler-dashboard-container {
+	.loading-container {
+		display: none;
+
+		&.is-active.is-loading {
+			display: block;
+		}
+	}
+}