--- a/client/performance-profiler/pages/loading-screen/index.tsx
+++ b/client/performance-profiler/pages/loading-screen/index.tsx
@@ -1,4 +1,3 @@
-import { ProgressBar } from '@automattic/components';
 import styled from '@emotion/styled';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useState } from 'react';
@@ -8,17 +7,38 @@ interface LoadingScreenProps {
 	isSavedReport: boolean;
 }
 
-const Progress = styled( ProgressBar )`
-	div {
-		background: linear-gradient( 90deg, #3858e9 0%, #349f4b 100% );
+const StyledLoadingScreen = styled.div`
+	padding: 15px 0;
+	h2 {
+		font-size: 20px;
+		font-weight: 500;
+		line-height: 26px;
+		margin-bottom: 30px;
+	}
+
+	span {
+		display: block;
+		font-size: 16px;
+		font-weight: 400;
+		line-height: 24px;
+		margin-bottom: 15px;
 	}
 `;
 
-const StyledLoadingScreen = styled.div`
-	padding: 100px 40px;
+const FootNote = styled.div`
+	max-width: 460px;
+	background-color: #f7f8fe;
+	padding: 25px;
+	margin-top: 65px;
 
-	h2 {
-		text-align: center;
+	h4 {
+		font-size: 14px;
+		font-weight: 500;
+		line-height: 20px;
+		margin-bottom: 10px;
+	}
+	p {
+		font-size: 14px;
 	}
 `;
 
@@ -28,30 +48,54 @@ export const LoadingScreen = ( { isSavedReport }: LoadingScreenProps ) => {
 	const progressHeadings = isSavedReport
 		? [ translate( 'Getting your report…' ) ]
 		: [
-				translate( "Crunching your site's numbers…" ),
-				translate( 'Analyzing speed metrics…' ),
-				translate( 'Comparing with top sites…' ),
-				translate( 'Finalizing your report…' ),
+				translate( 'Loading your site' ),
+				translate( 'Measuring Core Web Vitals' ),
+				translate( 'Taking screenshots' ),
+				translate( 'Fetching historic data' ),
+				translate( 'Identifying performance improvements' ),
+				translate( 'Finalizing your results' ),
 		  ];
 
-	const [ progress, setProgress ] = useState( 0 );
-	const [ tick, setTick ] = useState( 0 );
+	const [ step, setStep ] = useState( 0 );
 
 	useEffect( () => {
 		const timeoutId = setTimeout( () => {
-			const newProgress = 100 * ( 1 - Math.pow( 1.07, -tick ) );
-			setProgress( newProgress );
-			setTick( tick + 1 );
+			if ( step === progressHeadings.length - 1 ) {
+				return;
+			}
+			setStep( step + 1 );
 		}, 1000 );
 
 		return () => clearTimeout( timeoutId );
-	}, [ progress, tick ] );
+	} );
+
+	const stepStatus = ( index: number, step: number ) => {
+		if ( step > index ) {
+			return 'complete';
+		}
+		if ( step === index ) {
+			return 'running';
+		}
+		return 'incomplete';
+	};
 
 	return (
 		<LayoutBlock className="landing-page-header-block" width="medium">
 			<StyledLoadingScreen>
-				<h2>{ progressHeadings[ Math.floor( ( progress / 101 ) * progressHeadings.length ) ] }</h2>
-				<Progress value={ progress } total={ 100 } />
+				<h2>Testing your site’s speed…</h2>
+				{ progressHeadings.map( ( heading, index ) => (
+					<span key={ index } className={ stepStatus( index, step ) }>
+						{ heading }
+					</span>
+				) ) }
+				<FootNote>
+					<h4>{ translate( 'Performance Mattters' ) }</h4>
+					<p>
+						{ translate(
+							'Walmart found that for every one second improvement in page load time they achieved, conversions increased by 2%.'
+						) }
+					</p>
+				</FootNote>
 			</StyledLoadingScreen>
 		</LayoutBlock>
 	);