@WBerredo 2024-08-15
> The loading screen looks amazing! Thanks for the great work. I have minor feedback that we should tackle in the next iteration. So when hash already exists in the URL, we should not show the loading screen. Instead, we should show the loading report progress bar since we are not actually running the test.

I think we can tackle this in a follow up issue considering the test is being retrieved quickly enough now that is barely possible to read what is there.