@a8ci18n 2024-08-19
Translation for this Pull Request [has now been finished](https://translate.wordpress.com/deliverables/overview/16487105).